little refactor, added results on 50th epoch

+++ validation.py
@@ -27,7 +27,7 @@ iou_stats = torch.zeros(5)
 for i, (input, target) in tqdm(enumerate(val_data_loader)):
     
         pred, d1, d2, d3, d4, d5, d6 = model(input.to(device))
-        calculated_iou = iou(pred.detach().cpu(), target)
+        calculated_iou = iou(pred.detach().cpu().int(), target.int())
         iou_stats += calculated_iou
 
 print(f'IoU calculation has been finished.')
